Unify crypto implementation with Buy, remove leftovers.
