+++ .travis.yml
@@ -6,7 +6,6 @@ before_install:
   
 script:
   - set -o pipefail
-  - "./Scripts/test_crypto"
   - "./Scripts/test_buy"
   - "./Scripts/test_pay"
   - "./Scripts/test_sample"
