+++ Buy/Buy.h
@@ -31,3 +31,7 @@ FOUNDATION_EXPORT double BuyVersionNumber;
 
 //! Project version string for Buy.
 FOUNDATION_EXPORT const unsigned char BuyVersionString[];
+
+#ifndef COCOAPODS
+#import <Buy/MD5.h>
+#endif
