+++ Buy/Crypto/MD5.m
@@ -1,6 +1,6 @@
 //
-//  Crypto.h
-//  Crypto
+//  MD5.m
+//  Buy
 //
 //  Created by Shopify.
 //  Copyright (c) 2017 Shopify Inc. All rights reserved.
@@ -24,14 +24,39 @@
 //  THE SOFTWARE.
 //
 
-#import <UIKit/UIKit.h>
+#import "MD5.h"
+#import <CommonCrypto/CommonCrypto.h>
 
-//! Project version number for Crypto.
-FOUNDATION_EXPORT double CryptoVersionNumber;
+@implementation MD5
 
-//! Project version string for Crypto.
-FOUNDATION_EXPORT const unsigned char CryptoVersionString[];
++ (NSString *)data:(NSData *)data {
+    uint8_t digest[16];
+    [self data:data md5:digest];
+    
+    return [self string16FromDigest:digest];
+}
 
-#ifndef COCOAPODS
-#import <Crypto/Hash.h>
-#endif
++ (NSString *)string:(NSString *)string {
+    return [self data:[string dataUsingEncoding:NSUTF8StringEncoding]];
+}
+
++ (void)data:(NSData *)data md5:(uint8_t *)md5 {
+    __block CC_MD5_CTX context;
+    CC_MD5_Init(&context);
+    
+    [data enumerateByteRangesUsingBlock:^(const void *bytes, NSRange byteRange, BOOL *stop) {
+        CC_MD5_Update(&context, bytes, (CC_LONG)byteRange.length);
+    }];
+    
+    CC_MD5_Final(md5, &context);
+}
+
++ (NSString *)string16FromDigest:(uint8_t *)d {
+    return [NSString stringWithFormat:
+            @"%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x",
+            d[0],  d[1],  d[2],  d[3],  d[4],  d[5],  d[6],  d[7],
+            d[8],  d[9],  d[10], d[11], d[12], d[13], d[14], d[15]
+            ];
+}
+
+@end
