--- Scripts/test_crypto
@@ -1,11 +1,0 @@
- #!/usr/bin/env bash
-
-set -ex
-set -eo pipefail
-
-xcodebuild test \
--project "Buy.xcodeproj" \
--scheme "Crypto" \
--sdk iphonesimulator \
--destination 'platform=iOS Simulator,name=iPhone 7,OS=latest' \
- | xcpretty -c
